@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
 
 @Component({
   selector: 'app-root',
@@ -7,15 +6,7 @@ import { HttpClient } from '@angular/common/http';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  title = 'app';
-  result = '';
 
-  constructor(private http: HttpClient) {}
+  constructor() {}
 
-  private sayHello(): void {
-    this.result = 'loading...';
-    this.http.get('/api/hello-world').subscribe((response: any) => {
-      this.result = response.message;
-    });
-  }
 }
